[refactor] common/domain -> common/entity 로 패키지명 변경 - #33
