+++ novelservice/src/main/java/com/iucyh/novelservice/common/entity/DateAuditEntity.java
@@ -1,4 +1,4 @@
-package com.iucyh.novelservice.common.domain;
+package com.iucyh.novelservice.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
